char: oscar: fix -Wpointer-to-int-cast error
drivers/char/oscar/oscar_platform_driver.c:1065:18: error: cast to smaller integer type 'u32' (aka 'unsigned int') from 'void *' [-Werror,-Wvoid-pointer-to-int-cast]
        u32 intnc_val = (u32)data;
                        ^~~~~~~~~

Casting void* pointer into a 32bit length int data type raised
an error of `-Wvoid-pointer-to-int-cast`. Cast data into u64 first
then assign to u32 will auto truncate data without compiling errors.

Bug: 171827315

+++ drivers/char/oscar/oscar_platform_driver.c
@@ -1062,7 +1062,7 @@ static int oscar_lowprio_irq_notify(struct notifier_block *nb,
 	    container_of(nb, struct oscar_dev, lowprio_irq_nb);
 	struct gasket_interrupt_data *interrupt_data =
 		oscar_dev->gasket_dev->interrupt_data;
-	u32 intnc_val = (u32)data;
+	u32 intnc_val = (u64)data;
 
 	if (irq == ABC_MSI_AON_INTNC &&
 	    (intnc_val & 1 << INTNC_TPU_WIREINTERRUPT2))
